requirements.txt
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
pymysql==0.7.11
gunicorn==19.7.1
